Change Mesh.toCurve
This restores the default behavior of Mesh.toCurve, that was changed by
32dd145. It should not be assumed the the default behavior was
not useful for anybody. Furthermore, the docs state that it is
equivalent with Formex.toCurve.
The new behavior can be activated with an option (which may become
the default when the implementation is proven universal, the
docstring is corrected.