Reduce PolyLine coarsen methods
The bunch of PolyLine coarsen methods have been reduced to a single
one: an efficient implementation of Douglas-Peuckler decimation.
Having too many methods performing the same is not good: it is
confusing for the user and complicates maintenance of the software.
The current method does a very good job at keeping a good approximation
of the original PolyLine (depending on tolerance).

The new coarsen method has also provisions to force selected points
to be present in the coarsened result. This could e.g. be used to
set a preselection based on angle, curvature, or any other criterion.
There is (currently) no limit on the segment length: this could probably
be added as a general subdivide method.

The 'vtk' method was removed because VTK is not a prerequisite of
standard pyFormex, and thus vtk_itf should not be imported in other modules
(except those explictely dependent on it, like vmtk_itf).
The vtk method stays of course available through the vtk_itf module.

An example 'Polyline' has been added to experiment with the method
and the tolerance settings.