Fix/improve Curve/PolyLine.approx
Since the merging of approx and approximate, the equidistant approx
was broken. Fixed now.

PolyLine.approx is now equivalent with Curve.approx. This means
that PolyLine.approx now also has the 'chordal' method. By default
this will not coarsen the PolyLine, because the approx tries to
be as correct as possible, but an as yet undocumented feature
lets you now specify the initial approximation which could be
coarser than the input PolyLine.

atChordal should probably also get a maxlen parameter