Implement PolyLine.refine
PolyLine.refine is in one way the opposite of coarsen: it will add
points instead of removing points. But in another way it is the
companion of coarsen: it will make sure to get segments not longer
than maxlen.

The combination PL.coarsen(tol,maxlen).refine(maxlen) is the best
way to create a good aproximation for a PolyLine.
While the Curve.approx methods do work, they will not produce a
coarser PolyLine than the input, only a finer.
Now PL.coarsen wil create a coarser PL, keeping the points needed
for accuracy and keeping points to reduce the segment length as
needed. This may still leave segements longer than maxlen. Adding
.refine(maxlen) will ensure no segment is longer than maxlen.

Maybe we should add this combination as another approx method
besides 'chordal', 'parametric', 'equidistant'?