Added parameter return_indices to Mesh.reduceDegenerate
The function has been modified to work on the indices of the original mesh.
If return_indices is True a list of the indices of each Mesh is also returned.

Maybe the loop to check the position of the degenerate elements could be set in a
different fuction to that will work as Connectivity.testDegenerate but will
also discriminate the different elements degenerations.

A similar implementation could be used in Connectivity.reduceDegenerate,
and we could use the indices also to be returned from splitDegenerate.