Revert "Added parameter return_indices to Mesh.reduceDegenerate"
This reverts commit b581f6b533eb355e9813b02a14a459fa2b3d2e72.

The reason for reverting the commit is that we already have a
standard mechanism in pyFormex for getting the indices: the
prop attribute was meant especially for the purpose of finding
out the parent element for every generated geometry element.
So it is better to use that standard mechanism instead of
introducing ad hoc solutions.

Just set the prop attribute to a list of element numbers,
and after any transformation you find the parent element
number in the resulting prop:

  M.setProp(arange(M.nelems())
  redM = M.reduceDegenerate()

Note that we have some functions with the return_indices
parameter: these are functions that return non-Geometry
types, which do not have the prop attribute.