Revert "added parameter method in coords. sizes."
This reverts commit bdfc010d02355018eafb1b7833134fd6299b94d6.

Reverted: this functionality alread exists as
  F.principalBbox.sizes()

In general, to all developers: please refrain from adding new
functionality to the core modules, or making changes to it.
We need some basic functionalities that do not keep changing
all the time.

For the core modules, (in the pyformex directory), changes
should be limited to bug fixes and absolutely important feature
changes. The latter should preferably be proposed and discussed
first.

In the plugins directory you have more freedom to operate ;)