Add parts parameter to Mesh.fuse
The Mesh.fuse method now has an optional parameter parts.
If specified, it is a list with part number for each element
(such as e.g. the prop array) that is used to split the Mesh
in parts (see splitProp), and the the fusing is done part
by part. Elements that get a negative value will not be
involved in the fusing operations.