Add Coords.adjust method.
Coords.adjust is like Coords.fuse (and takes all the same parameters)
but instead of fusing closeby points it will just give them
identical coordinates. The shape of the Coords remains the same.

This method is also exposed to all Geometry types as a coords
transform. This can e.g. be useful in combination with the
the Mesh.fuse(parts) operation. See the FuseParts example.