Fix a bug in and improve submitJob
A bug in jobs_menu.submitJob is fixed, and it now has the option
to either transfer specified files, or all files in a job folder
(default).

As a result submitJob is now used in jobs_menu.submitToCluster.

The user interface of the jobs plugin 'Submit Abaqus Job' item
has been extended. (There's a lot more we could add)