Store image shape in drawImage3D
The drawimage3D now stores the image shape (width,height) in an
attribute _imageshape_ of the object. This is handy because the
image shape was lost in the Formex for the draw function.
Now, if FA is the actor returned by drawImage3D, you can retrieve
the image size as:
width,height = FA.object._imageshape_