Added IsoContour example. (bug #43977)
The ContourView example by Francesco gave us the idea to
implement a 2D equivalent of the 3D isosurface function (in the
isosurface plugin): isoline creates contour lines through
data in a 2D regular grid, at a given level value.
The 2D grid data can for example be pixel data from a
raster image.

The function uses a marching squares algorithm (the 2D version
of the marching cubes in isosurface). It is currently put in
an example (IsoContour), because we still need to create a
C version for efficiency (like we have for isosurface).
The Python version will later be moved to lib/misc.py.

The isoline function could directly be used on pixel data
from the rendering and thus form the basis for the suggestions
of Gianluca in bug #43977:
https://savannah.nongnu.org/bugs/?43977#comment1