Add accelerated isoline function (bug #43990)
A C-implementation of the marching square algorithm in the
isoline function was added to the lib.misc module.
A wrapper isoline function was added to plugins.isosurface.
This is the prefered way to call isoline: it will select the
accelerated function when available, or fall back to the
Python version if not.
Also, this function provides for multiprocessing, to even
further speed up. Default is currently set to nproc=1,
because multiprocessing isoline has not been tested yet.