Add QtCanvas.outline method
QtCanvas.outline returns the outline of the current rendering
as a plex-2 Formex. This works by rendering the current scene
in an offscreen buffer and grabbing the pixels in an array
(using QtCanvas.rgb). Then computing pixel luminance and
applying isosurface.isoline to get the contour in pixel
coordinates. Finally using the current camera transformation
matrices to transform the result back to world coordinates.
The resulting contour lies in a plane through the center of
the scene and parallel with the camera (not global!) axes.

The computation of the outline works in any orientation,
and in perspective as well as projective mode. Of course,
the outlines obtained in projective mode are the most useful.

See example ContourView for an application (and comparison
with the vtk implementation, which probably can be removed
after some teesting).

The drawing of the outline can also be initiated from
the Geometry menu.