Fixes for running pyFormex on Python3
As the Python world is moving forward, we are continuing
the work on porting pyFormex to Python3.
Here are some small fixes needed to run pyFormex on Python3.

Remember that we intend to the the porting in a compatible
way, i.e. the same code should run on both Python2 (2.6 or 2.7)
and Python3 (3.3 or higher).