Revert "Add parameter ref to inertia.inertia to select reference point"
This reverts commit 1551961248a0d398327cf05105bee60488603680.

The implementation was (in the expected case) incorrect, and the
idea is ambiguous.
There are three ways of transforming inertia from one system of
axes to a parallel one. The system through the mass center (G) always
has the lowest inertia, thus
- from G to X: I += tr
- from X to G: I -= tr
- from X1 to X2: I += tr2-tr1

The inertia function computes the mass center approximately. We do not
know if the user accepts this as G, and transforms to another point
(most likely, so should be + then) or wants to correct the results
because he knows the accurate G.

I suggest to create a general intertia transforming function like
inertia.translate(I,G,X,toG=False)