Add the old fe_abq interface as alternative
This commit add back the old fe_abq interface and provides
a startup option to select the old one instead of the default
new one.

The idea is to make the many bugfixes and improvements in the
master branch available to those following the branch 1.0.0
because they still require the old interface.

The divergence between the master and 1.0.0 branches was growing
too rapidly and supporting both branches became too cumbersome.

From now on, branch 1.0.0 is no longer supported. Users should
migrate to the master branch as soon as possible.

To use the old fe_abq from the master branch pyformex, do
  pyformex --oldabq.

The plugins.fe_abq.py module has become a wrapper module that
loads the definitions from fe_abq_old.py or fe_abq_new.py as
defined by the oldabq option.