Fix relative imports in freetype
The freetype bindings included in the pyformex source (this is now
standard) did not work, because of incorrect relative imports.
This error was not noticed because pyformex pocked up the previous
installation from the extras directory. But it would fail for a user
that had never installed freetype from extras.

Now the imports are fixed and only the modules inside the
pyformex/freetype path should be used. Older installations from the
extras can be removed.

Also, the release has been bumped to 1.0.1~a2.