Add Tensor and Inertia classes
Without much discussion, the Tensor2/Tensor3 debate in the
plugins.inertia module has been decided in favor of Tensor3,
which was renamed Tensor. It has only the rotation as operation.

A derived class Inertia can be used for an inertia tensor, and
has a translate method. The Inertia class stores the total mass
and center of mass. The center of mass can also be forced at
construction time, and the Inertia values will then be
corrected for it.

Todo:
- test (esp. the translate method)
- initialization of Inertia class from given inertia values.