Interactively set focus
The tools menu now has an option to interacively set the focus on
some part of the model.
- First you pick some points with Tools->Pick->Points.
- Then use Tools->With picked->Focus
The center of the picked points then becomes the focus of the
camera, while camera zooming remains (as much as possible) the same.
The selected point will thus be displayed in the middle of the screen,
and camera rotation operations will now evolve around the selected
point.

TODO: We should probably collect this functionality in a single
function, and binjd it to a toolbar button.