Added function gui.viewport.sanitizeSize
the functions sanitizes the suggested canvas sizes
None value will set the width (or heigth) to one the current canvas
A negative value will rescale the width (or heigth) will rescale the canvas
according to the non-negative value.
If both  width and heigth are negative, the canvas size are set as the
current canvas.

functions gui.viewport.image and gui.viewport.outline have been adjusted to the new function