Convert pyfFolder.xpm to pyformexdir.png
Icons with a lot of colors are better stored in .png format
that .xpm. We keep icons with a small number of colors in .xpm
because that is storage friendly and they can be easily hand-edited
pixel by pixel.

Also, icon names are all lower case, and we prefer pyformexdir, as
it is called inside pyFormex.

Now it would be good if someone can come up with a use for this
icon inside the pyFormex GUI.