Add configurable sidebar paths to FileDialog
widgets.FileSelection has been renamed to FileDialog. This is more
in line with the other Dialogs.

The FileDialog has got an extra parameter: sidebar. Here you can
specify a list of paths that will be added to the sidebar in the
FileDialogs. The default set of sidebar paths is set in the
configuration file and currently contains :
user's home directory, pyformex top level, pyformex data,
pyformex examples, system root (/).