Add inertia of volume inside surface
The TriSurface.inertia() methode computes the inertia properties
of the modeled *surface*. Often however, a TriSurface is used as a
model for the *volume* inside the (closed) surface.

There is now a TriSurface.volumeInertia() which computes the
inertia (and total mass and center of mass) of the volume inside
the surface, supposing a uniform density of 1.0 throughout the
volume.

For convenience, this method can also be called as
TriSurface.inertia(volume=True)

The geometry_menu->showPrincipal function will offer the user
the choice Volume or Surface. Try out both on e.g. the horse model
to see the impact.