Add octahedron element
elements.py now has an octahedron element. Not intended for base
of FE analysis, but as icosahedron can be used for geometrical
modeling, e.g. creating layouts for triangulated surfaces.

Used as example in simple.sphere, which has now an extra parameter
to set the layout based on icosahedron or octahedron.

geometry_menu->CreateObject->sphere has the option in the dialog.
Set ndiv=1 to see the base icosahedron/octahedron.