Improve utils.warning wrapper
The utils.warning wrapper will now update the functions metadata
if the level is UserWarning. That means that the function's
docstring is kept, so that it is still in the reference manual.

Higher warning levels, as. e.g. used in the utils.deprecated
wrapper, do not keep the docstring. So adding a deprecated*
decorator to a function still removes that function from
the reference manual.