Subclass inertia.Tensor from numpy.ndarray
The inertia.Tensor and inertia.Inertia classes are now subclasses
of numpy.ndarray. It is now easier to do computations and
manipulations with the tensor and its elements, without needing
to use the tensor attribute. The tensor attribute is still valid,
but will return the array as a ndarray instead of Tensor.