Make undraw work with tuples.
undraw(actors) did not work when a tuple of Actors was given.
Now you can specify a single actor, a list of actors or a tuple of
actors.

Deleting an Actor sometimes fails. It turns out that to find the
item to be deleted, a comparison is needed and comparison of a
complex actor may fail due to use of arrays.
A solution using a try..except has been implemented, and removal
by id when normal removal failed. We should probably generalize
the use of (Python or our own) ids for increased performance.