Allow newlines in drawText (fixes bug #43954)
textext.Text has been changed to process strings including
newlines. Thus drawText now supports multiline text, with 2D
as well as 3D positioning.

The input text is split on newlines and the first string is displayed
at the specified position. The following lines of text are then
created as child Text actors, with a position that is increased
vertically with a negative offset, resulting in positioning below
the previous line. The default offset is equal to the font height,
but a new parameter lineskip (default 1.0) allows you to specify
another value. Lineskip is relative, and can be set negative, in
which case each next line will go higher.