Improve CoordSys to replace CoordinateSystem
The CoordSys class has been improved so that it can now fully
replace CoordinateSystem, which is now deprecated.

The old CoordinateSystem class was derived from Coords, which
had the side effect that a lot of transformations were possible
that left the object in a non-valid state as coordinate system.

The new class is implemented as a rotation + translation matrix,
so that transformations using the CoordSys are readily available.
The CoordSys itself can only be transformed with a limited set
of functions (like rotate and translate) that keep its state
valid. We might add some more (like reflect, swapAxes) if there
is interest.