Changes to the inertia plugin
add parameter cs in Tensor to store the CoordSys stored in Tensor.CS
change Tensor.rotate to return a new tensor with CS stored as the original
cs roteted of the new value, Should the CS and the tensor furst be translated?
add Inertia.translateTo which still call the old trnslate (maybe to be merged)
add Inertia.toCS to rototranslate inertia to the new CS, should the
transalteTo method update the  translation of the CS of the Inertia?

dont know well how to deal with this things and how do you wanna deal with reference axes, to store the staic
mass moment