Cleanup imports of trisurface and inertia
After the move of trisurface and inertia, there were some bad
imports left (still importing from plugins). This had no big
consequences because the imports were not used, and most still
had a .pyc version of the module.

Now the .pyc module is cleaned up however, and the faulty imports
had to be removed.