FontTexture.texCoords now also accepts string
The FontTexture.texCoords now also accepts an ascii string as
input instead of a single character (ordinal number).
In case of a string, it returns an array of texture coordinates
for rendering of the whole text on a series of quad4 elements,
one character per element.

It is used in the Text class, and in the Mobius and TexText
examples.