Marked mesh_ext as deprecated.
mesh_ext.intersectionWithLines has been renamed to
quad4MeshIntersectionWithLines and is on longer a Mesh method.
This is because it is not correct for a quad4 Mesh (elements
can be non-flat) and it is thus better to let the user do the
approximation:

  p,i = M.toSurface().intersectionWithLines(...)

Note that over time I want to remove the mesh_ext completely.
If there is anything important there, it should be cleaned up
and moved into mesh or elsewhere.