Change geomtools.distances* functions (bug #41564)
distancesPFL and distancesPFS have been unified in a single
distanceFromLine function. The lines can eiter be specified
as a tuple (point,vector) or as an array with two points on
every line.

This will probably also be done for some other functions that have
a line and a segment variant.

A function distanceFromSegment can be added, but should
return the closest endpoint if projection of the point is
not inside the segment.