Adding pointNearLine2, because of memory usage
With the pointNearLine the memory explodes for large data.
A different implementation, similar to the closeLinesPoints included
in intersectionWithLines before commit ff98f52, has been added in
pointNearLine2.
The pointNearLine2 does not increase memory usage when data sizes increase
and it is faster. An example for testing has been added in bugs: #44422