Change the qt4 bindings setting
The default configuration variable 'gui/bindings' now holds 'any',
meaning that you do not have any preference for which bindings
('PySide' or 'PyQt4') to use. The system will the try to use PySide,
and if not available, try to use PyQt4, and if also not available,
fail.

You can change this to a fixed preference by changing the user prefs
(e.g through the Settings menu). pyFormex will then fail if the
required bindings are not available.

In any case, you can still override the configured setting by one of
the startup options:  --pyside  --pyqt4