Important changes in script mode (bug#44508)
Multiprocessing in script mode is now silently working when called
from inside a loaded module. A warning is only written when the
multitasking is started from inside the script (the change that it
works there is pretty small, maybe we should just warn and not
provide an option to continue).

The detection of being inside the script or inside a loaded module
was previously done by the test __name__ == 'draw'. This is now
considered unsafe, because it is easy for a user to have a
component in sys.path where a module draw.py lives (we already have
three draw.py modules in pyFormex alone), and functions in such
module would have __name__ == 'draw'. Therefore, during execution
in script mode, pyFormex now sets __name__ to '__script__' or
'__draw__', depending on being run in nongui or gui mode.

I know this will cause troubles for existing scripts, but here are
some guidelines:

- if you want to use the latest pyformex only, change
    if __name__ == 'draw':
  to
    if __name__ == '__draw__':
  You can put the following near the top to make sure that you only
  run on a recent enough pyformex:

    utils.requireModule('pyformex','1.0.2~a1')

- if you want to use your script with both new and older versions,
  you can test for script execution as follows:

    from distutils.version import LooseVersion
    if __name__ == '__draw__' or ( \
       __name__ == 'draw' and \
       LooseVersion(pf.__version__) < LooseVersion('1.0.2') ):

pyFormex version has been bumped to '1.0.2~a1' to mark this change.