Revert "fixed small bug in properties.checkArrayOrIdValue"
This reverts commit a3e94c152ebbc2bd6e87f4b508f359cdc28ab19e.

The commit claimed to fix a bug, but it rather introduced new bugs.
The included examples did not even run anymore:
  pyformex --testmodule plugins.properties

Devs: please refrain from making bold changes like this!
The proper way of handling the calimed bug would be to file a bug report,
add an example of the failing function, suggest a fix, discuss, fix
and finally test.
Otherwise we are just continuously introducing new bugs with every
change, and stable and reliable code will always be far off.

And run at least the tetmodule command on the changed modules.