Do not wait for browser commands (bug #44821)
Added 'wait=False' to the execution of the browser command when
displaying help. This avoids blocking pyFormex until the browser
command exits, if the browser was spawned as a child.
If the browser was started before pyFormex, there was no problem,
obviously. Now you can also let pyFormex start the browser when
needed.

The 'wait=False' has also been added to the fileMenu.recordSession,
but that case has not been tested yet.