Fixed a bug in Curve.atApproximate, Curve.chordal
For closed curves the Curve.atApproximate erroneously stripped off
the last point, making the Curve.chordal approximation incorrect
in the last segment of a closed curve.
Previously we didn't store the last=first point, but now the
BezierSpline *does*, so the last point should be included.

Note that a closed PolyLine *does not* store the last point, so you
might get an extra point here. However, the choral approximation of
a PolyLine is obviously itself, so no problem arises.