Modified fe_abq_old.exportMesh to save different sets per each prop set
Function works as before but creates also different sets if
multiple prop are stored. The parameter eltype has no default anymore
and must be set explicitely. Previously the in eltype==None
the abaqus eltype was set equal to the pyformex eltype name.
There is no point in this as the name differs and abqus will raise error
anyway.