Change version numbers of alpha releases
From now on will (again) use {VERSION}-a{BUILD} as the release number
for an alpha release, using a dash instead of the tilde character.
The reason is that the dash is lexically sorted before the dot, while
the tilde is after it. Thus, in a directory listing with some alpha
releases and final releases, the alpha releases are sorted before the
final release. This avoids people selecting the alpha release thinking
it is a later version.

This versioning schedule was also used before, but we changed it to
tilde when we started creating Debian packages. The Debian package
versioning does not allow the dash, while it interpretes the tilde
before the final version. But since we normally do not create Debian
packages for alpha versions, that should not give a problem. And in
case we want to create alpha Debian packages, we should be able to
rename the files in the Debian build procedure.