Change TriSurface.nonManifoldEdges
TriSurface.nonManifoldEdges has been renamed to nonManifoldEdgesFaces,
as it returns both edges and faces.
A new method TriSurface.nonManifoldEdges has been added that only
returns the edges. This is to conform with Mesh.nonManifoldEdges
which also just returns the edges.
In future, we might include TriSurface.nonManifoldEdges into the
Mesh method to have it work for all level 2 Meshes (the current
only supports level 3 Meshes).