Improve selection PySide/PtQt4
The default is now (almost) neutral and will use what is available
with first choice being PySide if both are.
If a bindings is set to PySide or PyQt4, then only that one will
be used.
If no success in loading the requested one, or any if unspecified,
an exception is raised.