Enforcing coding style recommendations
In an effort to create better, more understandable and maintainable
code, we will raise the requirements on the produced code.
In a first step we advise to check the produced code with
available tools (pep8, pylint) and over time more stringent
rules will be set.

It is good to get used to these tools already.
See the HOWTO-dev for more info on how to use.

simple.py and varray.py have been treated and improved as examples:
- simple.py only checked with pep8
- varray.py checked with pep8, then pylint