Add smart PolyLine concatenation
PolyLine.append has now an optional `smart` argument. If set True,
the procedure will reverse the appended PolyLine if its last point
is closer to the end of the current PolyLine than its first point.

There is also a new (static) PolyLine.concatenate function which
directly concatenates a whole list of PolyLines. Arguments are like
in PolyLine.append.

Finally, the new example PolyLineConcat illustratest the use of the
new concatenate function and the effect of the `smart` argument.