Create separate module for Process class
The Process class can conveniently be used to execute external
commands in other projects. To facilitate such use, it is split
off from the utils module and put in a separate module 'process'.

Then on any system with pyfomrex installed you can just do

  from pyformex.process import Process
  P = Process('my_external_command',...)

with all the extras delivered by the Process class.