fix in gui Tools/With picked/Set property
The Set property failed before this change.
The Set property calls the setpropCollection which
at the end needs to redraw the
actor with the new colors taken from the prop numbers.
The redraw() does not exists. In this commit it has been
replaced by first removing and then addeding the actor.
Pobably, the combination of removeActor and addActor
could be written in a separate function redrawActor()