Rename default settings file to pyformexrc
Some time ago the default settings file was renamed from pyformexrc
to pyformex.conf. This create a conflict however with the user
settings files, because a 'pyformex.conf' file in the directory
from where you start pyformex is automatically recognized as the
user settings file, and used to store user changes to the settings.
As a consequence, when starting pyformex from the source directory
holding the pyfomrex executable (and the default settings), the
default settings got overwritten.

This commit restores the old 'pyformexrc' name of the default
settings.