extend tools menu with 2D measurements: point 2D and angle 2D
In addition to the 2D distance now the user can query the
point in 2D (the horizontal and vertical distance from origin, defined
along the camera axes), and the angle in 2D (the angle using the camera
focus-eye axis). The advantage of these functionalities is
that you do not need to pick existing 3D points, but you can click
anywhere on the screen.