Use the filename InputItem wherever possible.
Some uses of a general 'button' type widgets were replaced with
the new specialized button for filenames.
Places where the filename referes to an image and is accompanied
by a preview widget have not been changed yet. Also buttons aimed
at the input of a directory path have not been changed yet.
We have to either create specialized itemtypes for these, or
add special optional arguments.