Add image preview to filename InputItem
The filename InputItem now has an extra argument preview= to set
a widget for previewing the selected image. Obviously you should
use a fileter='img' (or more narrow, e.g. 'png' or 'jpg'...)
to make that work.

The new functionality is used in the settings for the splash image,
in the viewport menufor the background image, and in the
SurfaceProjection example.

It is however not used in some other cases where input of an image
file name is done, because they require a customized function
(like showing the image size). Maybe we could add that later to
a standard image preview widget.