Fix a bug in Coords.projectOnSurface
An important bug was lingering in Coords.projectOnSurface
(since May 2015). Commit b74c8b125 change the behavior of
Coords.closestToPoint (returning an index instead of point
itself). But the change was not accounted for in projectOnSurface.

This should warn us that we need to add more tests to the
source code (preferably doctest, as they go into the
documentation).