Fix a bug in Attributes
Now an attribute given a value None is removed from the Attributes
object independent of the method used for setting the value.
Thus all three of the following will remove an item with key 'color'
from the attrib object:
  attrib(color=None)
  attrib.color = None
  attrib['color'] = None

Because the default behavior of Attributes is to return None for
any missing key, we expect this will not cause any misbehavior.

But in cases where explicitely was tested on the existence of a key,
a change in behavior may result. So please report suspected
malfunctioning.