temporary fix for contact interaction for abaqus
On 04 Feb 2016 a commit (1b948d) changed the interaction from dictionary
to Interaction. This change is premature as it breakes many functionalities
of the contact (e.g.  'contacttype':'SURFACE TO SURFACE', 'pressureoverclosure':...).
Moreover, class Interaction is work in progress and not well documented.
With this commit both the old stable behaviour (not using the class Interaction)
and the new implementation (using the class Interaction) are supported.
In future, the contact implementation should be reviewed.