Improve autoglobals feature.
Autoexport form a script/app is now controlled by a single global
variable, autoglobals, which can be set to:
   False | True |   [ NAME | CLASS ]
The autoclasses variable is no longer used.
Also, a list of the autoexported globals is printed when finishing
the script.
And invalid entries are silently ignored, to make the execution of
the script finish properly.