Give Actor an attribute pickable (#42075)
An Actor now has an attribute pickable, which is True by default
and can be set to False to disable picking operations on that Actor.
A pf.canvas.pickable list can still be set to force a list of
pickable actors other than the one defined by the
Actor.pickable attributes.
In all cases, the picking operations now return the Actor index
from the pf.canvas.actors list. The pf.canvas.pickable is intended
for internal use in a specialized picking function.