Add Curve.position and Curve.sweep
Curve.position and Curve.sweep perform similar actions as (and can
in most cases replace) coords.postionCoordsObj, coords.sweepCoords
and Mesh.sweep.

The new functions are better than the old ones, in that they are
using the (moving) frenet frames along the curve to place the
objects without risk of kinking, but they are missing (for now?)
the scaling available in coords.sweepCoords.

Because of the use of the Curve.frenet() method, it was more natural
to put them in Curve instead of in Mesh. This also provides a smooth
upgrade path. Instead of aMesh.sweep(aCurve), use aCurve.sweep(aMesh)
to have the new functionality.

TODO:
- implement csys parameter (replacing normal and upvector).
- implement approximation parameters to be passed to frenet?
  (users can just create a Curve approximation first)
- implement the scaling of the object along the curve?
  (how: maybe a general trf matrix at all curve points?)

Example: the Curves example showcases the new feature. Choose e.g.
the data set 2 with BezierSpline, and check the Sweep option.