Make the axes triade customizable.
In the gl2 engine, the axes triade was hardwired to a set of
three lines along the axes.
Now the displayed triade respects the coordinates of the
decors.Triade Actor, allowing for refinement of the triade
as was available with the gl1 engine.

Also, the Canvas.setTriade has got an extra parameter 'triade'
allowing to specify another object to be displayed instead of
the default triade. THus the use can fully customize it.

A 'triade' pattern was added in simple.Pattern.

Some unneeded legacy code was removed.