Fix the drawing of principal axes.
The drawing of the principal axes from the geometry menu was
incorrect. It performed a transpose on the rotation matrix
when constructing a CoordSys. But the transposition is already
included in the CoordSys when initializing with points, and
should not be done when initializing with a rotation matrix
(which has the unit vectors in columns).
Also, the inertia.principal docs have been amended to clear
this up.

The drawAxes and drawPrincipal functions have been reworked
to use the new gl2 engine.

Furthermore, this commit removes all legacy Actors, except
for the NurbsActor, which is the only thing to be converted.
(Quadratic drawing is also missing, but has already been for
some time. Maybe we can some day revive it together with
Nurbs).