Add a new color type to Actors
The color (and bkcolor) of an Actor can now be set by a string
of the form 'fld:field_name'. In that case the color will be taken
from a field named 'field_name' stored in the object that is
represented by the Actor. The field should be of type 'node'
and specify the RGB colors of all nodes.

Also, Drawable has a new method, changeVertexColor.
It can change the color of an already drawn Actor.
Beware, this is experimental, and only works for specific
cases (vertexColor on front faces of drawable[0]).