Add new class imagearray.DicomStack
The DicomStack class collects a stack of DICOM images in a single
object. It provides extended functionality over the dicom2numpy
function.
One of the benefits is that it reads the z-position of the images
and sort the image files accordingly.
Currently it only works with python-gdcm, not with python-dicom.