Add the check item back to MessageBox (bug #48205)
The check item was removed in f94c1744c5c0bc, in preparation
of a replacement of MessageBox with a generic InputDialog.
It was thought that it was not used anyhow.
The transformation to using InputDialog has not been finished,
and the check item proved to be used, therefore it is
reinstated here.