Add new candy module; remove AxesActor
A new pyFormex module has been added to the core: candy.
It will be a collection of ready-made geometries that are more
complex than the ones in module simple.
I hope you will see the reasoning for the name:
ready-made -> (FR)confection -> (EN)confection: treat, candy.

A first object in the candy module is the Axes class: it
represents the axes (and planes) of a coordinate system.
It is the geometry that was used in the AxesActor before.
The AxesActor has been removed, as you can now just draw
the Axes. The drawAxes function is retained but does just
this.