extended Mesh.subdivide to non-linear elements
By taking advantage of the isopar.py the subdivide can
be applied to non-linear elements. For the moment, the code
has been included in mesh_ext.py, but in future the subdivideIsopar
should be the only function exposed to the user.
The current Mesh.subdivide is handled by subdivideIsopar as a
specific case.
The subdivideIsopar works with line2, line3,tri3 and higher orders,
quad4 and higher orders, hex8 and higher orders.
It should work also on tet4 and higher orders but the Mesh.subdivide
does not define this subdivision yet.
An example of the new subdivideIsopar will be included in Savannah.