Fix some bugs and bad practices.
Fixes were suggested while I was testing a new (for me) editor: spyder.
It is Python specific, written in Python and provides a lot of
Python functionality. A bit like Eric, but less overkill.
It continuously flags (possibly) bad statements, like undefined
variables and unused imports. It even flags all the places with
a TODO comment. Very handy.
I am missing the convenience of indenting and searching from Emacs,
but the source check functionality is enough to make me try it out
for some time.
If I can get used to it, I might one day integrate it into pyFormex :)