Compare Actors by id instead of contents (fixes bug #48240)
The base class for Actors (which was also renamed from Base
to BaseActor) now overrides the comparison operator __eq__.
It will make A == B equivalent with A is B.

This allows reliable (and fast) checking whether an Actor
is in a list. Basically, it compares by instead of by
contents. This makes sense as the Actor can contain very
complex attributes.

Note that because of this, when B is a copy of A, then
A == B will be False! But as there is no reason to make copies
of Actors, this would not hurt.