Rework the Triade customization. (Fix bug #48008)
The customization of the Triade has been cleaned up and made a
lot easier. The setTriade function now accepts any Geometry
as triade argument, as well as a string describing one of the
predefined triades. Currently available are 'axes' for the default
Triade, and 'man' for a realistic model of Gianluca ;)
The 'man' model is stored in data/man.pgf. The format allows to
store display attributes (only color is currently read back
however).
Using a Geometry as triade argument allows setting any drawing
attributes.