Store Geometry.attrib in PGF format files.
The PGF format has been upgrade from version 1.9 to 2.0.
In the new version, the .attrib dict of a Geometry
will also be stored in the .pgf file, (and read back).
The feature is experimental: it works ok with simple
attributes (string, single variable, list), but has not
been tested yet with more complex attributes.
Therefore we have retained an option to still write using
version 1.9. The default is version 2.0, but can read
older format. Older readers can read 2.0 but will ignore
the new feture.

The man.pgf file has been upgraded to version 2.0, and
contains a lighting=True setting, giving a better default
view of the 'man' Triade.